[REALESTATE-21] 수량 long 변경

+++ src/main/java/com/labshigh/realestate/internal/api/item/dao/ItemDao.java
@@ -35,8 +35,8 @@ public class ItemDao {
   private String imageUri;
   private String projectName;
   private String totalPrice;
-  private int quantity;
-  private int currentQuantity;
+  private long quantity;
+  private long currentQuantity;
   private int allocationDay;
   private String right;
   private String location;
